fix: addressed comments
